backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "township-backend"
version = "0.1.0"
description = "Township Request Management System backend"
authors = [{name = "West Windsor Forward"}]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",
    "httpx>=0.26.0",
    "tenacity>=8.2.3",
    "python-jose>=3.3.0",
    "cryptography>=42.0.5",
    "jinja2>=3.1.3",
    "reportlab>=4.1.0",
    "shapely>=2.0.2",
    "geojson>=3.1.0",
    "google-cloud-aiplatform>=1.48.0",
    "python-dateutil>=2.9.0",
    "pypdf>=4.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx[cli]>=0.26.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "pre-commit>=3.6.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
